Add test for LogModelResult
